+++ Assets/PatchKit Patcher/Scripts/AppData/Local/ZipUnarchiver.cs
@@ -1,4 +1,5 @@
-﻿using Ionic.Zip;
+﻿using System.IO;
+using Ionic.Zip;
 using PatchKit.Unity.Patcher.Debug;
 using PatchKit.Unity.Patcher.Cancellation;
 
@@ -16,6 +17,7 @@ namespace PatchKit.Unity.Patcher.AppData.Local
         private readonly string _password;
 
         private bool _unarchiveHasBeenCalled;
+        private MapHashExtractedFiles _mapHashExtractedFiles;
 
         public event UnarchiveProgressChangedHandler UnarchiveProgressChanged;
 
@@ -25,10 +27,11 @@ namespace PatchKit.Unity.Patcher.AppData.Local
         // not used
         public bool HasErrors { get; private set; }
 
-        public ZipUnarchiver(string packagePath, string destinationDirPath, string password = null)
+        public ZipUnarchiver(string packagePath, string destinationDirPath, MapHashExtractedFiles mapHashExtractedFiles, string password = null)
         {
             Checks.ArgumentFileExists(packagePath, "packagePath");
             Checks.ArgumentDirectoryExists(destinationDirPath, "destinationDirPath");
+            Assert.IsNotNull(mapHashExtractedFiles);
 
             DebugLogger.LogConstructor();
             DebugLogger.LogVariable(packagePath, "packagePath");
@@ -37,6 +40,7 @@ namespace PatchKit.Unity.Patcher.AppData.Local
             _packagePath = packagePath;
             _destinationDirPath = destinationDirPath;
             _password = password;
+            _mapHashExtractedFiles = mapHashExtractedFiles;
         }
 
         public void Unarchive(CancellationToken cancellationToken)
@@ -69,8 +73,13 @@ namespace PatchKit.Unity.Patcher.AppData.Local
         private void UnarchiveEntry(ZipEntry zipEntry)
         {
             DebugLogger.Log(string.Format("Unarchiving entry {0}", zipEntry.FileName));
-
-            zipEntry.Extract(_destinationDirPath, ExtractExistingFileAction.OverwriteSilently);
+            MemoryStream memoryStream = new MemoryStream();
+            string destPath = Path.Combine(_destinationDirPath, _mapHashExtractedFiles.Add(zipEntry.FileName));
+            zipEntry.Extract(memoryStream);
+            using (var target = new FileStream(destPath, FileMode.Create))
+            {
+                memoryStream.WriteTo(target);
+            }
         }
 
         protected virtual void OnUnarchiveProgressChanged(string name, bool isFile, int entry, int amount, double entryProgress)